printed floats

+++ MCU/lib/helper.h
@@ -2,12 +2,12 @@
 #define HELPER_H
 
 struct imu_values {
-  int x_acc;
-  int y_acc;
-  int z_acc;
-  int x_rot;
-  int y_rot;
-  int z_rot;
+  float x_acc;
+  float y_acc;
+  float z_acc;
+  float x_rot;
+  float y_rot;
+  float z_rot;
 };
